+++ index.js
@@ -1,8 +1,8 @@
 const maybeAddUnit = (value, unit) =>
   value + (typeof value !== 'number' || value === 0 ? '' : unit);
 
-const maybePixels = value => maybeAddUnit(value, 'px');
-const maybeDegrees = value => maybeAddUnit(value, 'deg');
+const pixels = value => maybeAddUnit(value, 'px');
+const degrees = value => maybeAddUnit(value, 'deg');
 
 // If passed an array, returns that array.
 // If passed a non-array, returns an array containing that value.
@@ -13,35 +13,42 @@ const stringifyList = (list, unit) =>
     .map(value => maybeAddUnit(value, unit))
     .join(', ');
 
-const maybePixelsList = values => stringifyList(values, 'px');
-const maybeDegreesList = values => stringifyList(values, 'deg');
+const pixelsList = values => stringifyList(values, 'px');
+const degreesList = values => stringifyList(values, 'deg');
 const unitlessList = values => stringifyList(values, '');
 
-const stringifiers = {
-  x: v => `translateX(${maybePixels(v)})`,
-  y: v => `translateY(${maybePixels(v)})`,
-  z: v => `translateZ(${maybePixels(v)})`,
-  translate: v => `translate(${maybePixelsList(v)})`,
-  translate3d: v => `translate3d(${maybePixelsList(v)})`,
-  scale: v => `scale(${unitlessList(v)})`,
-  scaleX: v => `scaleX(${v})`,
-  scaleY: v => `scaleY(${v})`,
-  scaleZ: v => `scaleZ(${v})`,
-  scale3d: v => `scale3d(${unitlessList(v)})`,
-  rotate: v => `rotate(${maybeDegrees(v)})`,
-  rotateX: v => `rotateX(${maybeDegrees(v)})`,
-  rotateY: v => `rotateY(${maybeDegrees(v)})`,
-  rotateZ: v => `rotateZ(${maybeDegrees(v)})`,
-  skew: v => `skew(${maybeDegreesList(v)})`,
-  skewX: v => `skewX(${maybeDegrees(v)})`,
-  skewY: v => `skewY(${maybeDegrees(v)})`,
-  perspective: v => `perspective(${maybePixels(v)})`
+const stringifiers = addUnits => {
+  const maybeDegrees = addUnits ? degrees : v => v;
+  const maybeDegreesList = addUnits ? degreesList : unitlessList;
+  const maybePixels = addUnits ? pixels : v => v;
+  const maybePixelsList = addUnits ? pixelsList : unitlessList;
+
+  return {
+    x: v => `translateX(${maybePixels(v)})`,
+    y: v => `translateY(${maybePixels(v)})`,
+    z: v => `translateZ(${maybePixels(v)})`,
+    translate: v => `translate(${maybePixelsList(v)})`,
+    translate3d: v => `translate3d(${maybePixelsList(v)})`,
+    scale: v => `scale(${unitlessList(v)})`,
+    scaleX: v => `scaleX(${v})`,
+    scaleY: v => `scaleY(${v})`,
+    scaleZ: v => `scaleZ(${v})`,
+    scale3d: v => `scale3d(${unitlessList(v)})`,
+    rotate: v => `rotate(${maybeDegrees(v)})`,
+    rotateX: v => `rotateX(${maybeDegrees(v)})`,
+    rotateY: v => `rotateY(${maybeDegrees(v)})`,
+    rotateZ: v => `rotateZ(${maybeDegrees(v)})`,
+    skew: v => `skew(${maybeDegreesList(v)})`,
+    skewX: v => `skewX(${maybeDegrees(v)})`,
+    skewY: v => `skewY(${maybeDegrees(v)})`,
+    perspective: v => `perspective(${maybePixels(v)})`
+  };
 };
 
-function transform(transformProperties = {}) {
+function transform(transformProperties = {}, addUnits) {
   return Object.entries(transformProperties)
     .map(([name, value]) => {
-      const stringifier = stringifiers[name];
+      const stringifier = stringifiers(addUnits)[name];
 
       if (!stringifier) {
         throw new Error(`Property '${name}' is not supported`);
@@ -52,4 +59,7 @@ function transform(transformProperties = {}) {
     .join(' ');
 }
 
-export default transform;
+export default {
+  transform: transformProperties => transform(transformProperties, true),
+  transformUnitless: transformProperties => transform(transformProperties)
+};