+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 1.0.0
 
 * Removes transpilation. Users must now transpile ES6 if needed.
-* Adds support for `skewX` and `skewY`
-* Adds default value for `transform` argument
+* Now exports an object instead of a function.
+* Adds `transformUnitless`
+* Adds support for `skewX` and `skewY`.
+* Adds default value for `transform` argument.